# artifacts/api-server/config/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ride-Hailing & Delivery Platform API",
    "description": "API specification for the Go (Chi/Gin) modular monolith ride-hailing platform in Rwanda & East Africa.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    },
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local Development Server, API v1 base path"
    }
  ],
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register or Request Login OTP",
        "description": "Registers a new phone number or triggers an OTP login for an existing number. Capture device information for collision audits.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phone_number", "device_id", "platform"],
                "properties": {
                  "phone_number": { "type": "string", "example": "+250780000000" },
                  "device_id": { "type": "string", "example": "device_iphone15_pro" },
                  "platform": { "type": "string", "example": "ios" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OTP sent successfully" }
        }
      }
    },
    "/api/v1/auth/verify-otp": {
      "post": {
        "tags": ["Auth"],
        "summary": "Verify OTP and Receive Token",
        "description": "Exchanges an SMS OTP code for access and refresh tokens. Role state indicates onboarding progress.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phone_number", "otp"],
                "properties": {
                  "phone_number": { "type": "string", "example": "+250780000000" },
                  "otp": { "type": "string", "example": "123456" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": { "type": "string" },
                    "refresh_token": { "type": "string" },
                    "role_state": { "type": "string", "example": "CUSTOMER_ONLY" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh Access Token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refresh_token"],
                "properties": {
                  "refresh_token": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New access token granted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log out from Platform",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "204": { "description": "Successfully logged out" }
        }
      }
    },
    "/api/v1/customer/profile": {
      "get": {
        "tags": ["Customer"],
        "summary": "Get Profile details",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Profile details returned" }
        }
      },
      "put": {
        "tags": ["Customer"],
        "summary": "Update Profile details",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "full_name": { "type": "string", "example": "John Doe" },
                  "fcm_token": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Profile updated successfully" }
        }
      }
    },
    "/api/v1/customer/location": {
      "post": {
        "tags": ["Customer"],
        "summary": "Retrieve Anonymized Nearby Drivers",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["lat", "lng"],
                "properties": {
                  "lat": { "type": "number", "format": "double", "example": -1.9441 },
                  "lng": { "type": "number", "format": "double", "example": 30.0619 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "List of nearby driver vehicle types and distance returned" }
        }
      }
    },
    "/api/v1/customer/rides": {
      "post": {
        "tags": ["Customer"],
        "summary": "Book a Ride Request",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["pickup_lat", "pickup_lng", "pickup_address", "dest_lat", "dest_lng", "dest_address", "transport_type"],
                "properties": {
                  "pickup_lat": { "type": "number", "format": "double", "example": -1.9441 },
                  "pickup_lng": { "type": "number", "format": "double", "example": 30.0619 },
                  "pickup_address": { "type": "string", "example": "Kigali Heights, Kigali" },
                  "dest_lat": { "type": "number", "format": "double", "example": -1.9582 },
                  "dest_lng": { "type": "number", "format": "double", "example": 30.0912 },
                  "dest_address": { "type": "string", "example": "Kigali International Airport" },
                  "transport_type": { "type": "string", "enum": ["MOTO_BIKE", "CAB_TAXI", "HEAVY_FUSO", "LIGHT_HILUX"], "example": "MOTO_BIKE" },
                  "initial_fare": { "type": "number", "format": "double", "nullable": true, "description": "Optional customer-provided starting fare. The API stores this value for negotiation data; it is not a system-generated fare suggestion.", "example": 1500.00 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Ride booked, search sequence started" }
        }
      },
      "get": {
        "tags": ["Customer"],
        "summary": "List Previous Rides",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Paginated ride history list" }
        }
      }
    },
    "/api/v1/customer/rides/{ride_id}": {
      "get": {
        "tags": ["Customer"],
        "summary": "Get Specific Ride details",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "ride_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Ride detail returned" }
        }
      },
      "delete": {
        "tags": ["Customer"],
        "summary": "Cancel Pending Ride Request",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "ride_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Ride successfully cancelled" }
        }
      }
    },
    "/api/v1/customer/rides/{ride_id}/negotiation/propose": {
      "post": {
        "tags": ["Customer"],
        "summary": "Customer Counter-offer",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "ride_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount"],
                "properties": {
                  "amount": { "type": "number", "format": "double", "example": 1500.00 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Offer sent to driver" }
        }
      }
    },
    "/api/v1/customer/rides/{ride_id}/negotiation/accept": {
      "post": {
        "tags": ["Customer"],
        "summary": "Customer Accepts Driver's Fare",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "ride_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Fare confirmed and locked" }
        }
      }
    },
    "/api/v1/customer/rides/{ride_id}/negotiation/decline": {
      "post": {
        "tags": ["Customer"],
        "summary": "Customer Declines Driver's Fare",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "ride_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Offer declined" }
        }
      }
    },
    "/api/v1/driver/apply": {
      "post": {
        "tags": ["Driver Onboarding"],
        "summary": "Submit Driver Profile Application",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["transport_type", "vehicle_plate", "license_number", "date_of_birth", "city", "momo_pay_code"],
                "properties": {
                  "transport_type": { "type": "string", "enum": ["MOTO_BIKE", "CAB_TAXI", "HEAVY_FUSO", "LIGHT_HILUX"], "example": "MOTO_BIKE" },
                  "vehicle_plate": { "type": "string", "example": "RAA123B" },
                  "license_number": { "type": "string", "example": "DL-987654321" },
                  "date_of_birth": { "type": "string", "format": "date", "example": "1995-05-15" },
                  "city": { "type": "string", "example": "Kigali" },
                  "momo_pay_code": { "type": "string", "example": "654321" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Application submitted successfully" }
        }
      }
    },
    "/api/v1/driver/profile": {
      "get": {
        "tags": ["Driver"],
        "summary": "Get Driver Profile",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Profile details returned" }
        }
      }
    },
    "/api/v1/driver/availability": {
      "post": {
        "tags": ["Driver"],
        "summary": "Set Online/Offline Status",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["is_online"],
                "properties": {
                  "is_online": { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Availability set" }
        }
      }
    },
    "/api/v1/driver/location": {
      "post": {
        "tags": ["Driver"],
        "summary": "Update GPS Location",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["lat", "lng", "speed_kmh", "heading"],
                "properties": {
                  "lat": { "type": "number", "format": "double", "example": -1.9448 },
                  "lng": { "type": "number", "format": "double", "example": 30.0625 },
                  "speed_kmh": { "type": "number", "format": "double", "example": 35.5 },
                  "heading": { "type": "number", "format": "double", "example": 180.0 }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Location captured, verified, and indexed" }
        }
      }
    },
    "/api/v1/driver/rides/{ride_id}/accept": {
      "post": {
        "tags": ["Driver"],
        "summary": "Accept Ride Request",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "ride_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Request accepted" }
        }
      }
    },
    "/api/v1/driver/rides/{ride_id}/decline": {
      "post": {
        "tags": ["Driver"],
        "summary": "Decline Ride Request",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "ride_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Request declined" }
        }
      }
    },
    "/api/v1/driver/rides/{ride_id}/cancel": {
      "post": {
        "tags": ["Driver"],
        "summary": "Cancel After Pickup Expiry",
        "description": "Driver-only cancel path for customer no-show. Allowed only after pickup_expired=true, and does not record a driver decline penalty.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "ride_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Ride cancelled as customer no-show" }
        }
      }
    },
    "/api/v1/driver/rides/{ride_id}/negotiation/propose": {
      "post": {
        "tags": ["Driver"],
        "summary": "Driver Counter-offer",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "ride_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount"],
                "properties": {
                  "amount": { "type": "number", "format": "double", "example": 1800.00 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Offer sent to customer" }
        }
      }
    },
    "/api/v1/driver/rides/{ride_id}/negotiation/accept": {
      "post": {
        "tags": ["Driver"],
        "summary": "Driver Accepts Customer's Fare",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "ride_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Fare locked" }
        }
      }
    },
    "/api/v1/driver/rides/{ride_id}/negotiation/decline": {
      "post": {
        "tags": ["Driver"],
        "summary": "Driver Declines Customer's Fare",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "ride_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Offer declined" }
        }
      }
    },
    "/api/v1/driver/rides/{ride_id}/negotiation/lock-fare": {
      "post": {
        "tags": ["Driver"],
        "summary": "Lock Manually Agreed Fare",
        "description": "Driver-only MVP endpoint for verbal fare agreements. It locks agreed_fare and confirms the ride without consuming negotiation offer count.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "ride_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount"],
                "properties": {
                  "amount": { "type": "number", "format": "double", "example": 2000.00 }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Fare locked and ride confirmed" }
        }
      }
    },
    "/api/v1/driver/rides/{ride_id}/negotiation/initiate-call": {
      "post": {
        "tags": ["Driver"],
        "summary": "Initiate Phone Call Bidding",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "ride_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Call logging initialized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "masked_number": { "type": "string", "example": "+250000000000" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/driver/stats": {
      "get": {
        "tags": ["Driver"],
        "summary": "Get Performance Stats",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Decline, acceptance, and priority tier metrics" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste your JWT access_token here."
      }
    }
  }
}